test-requirements.txt
ansible==7.6.0
ansible-lint==6.16.2
yamllint==1.32.0
flake8==6.0.0
pytest-testinfra==8.1.0
molecule==5.0.1
molecule-docker==2.1.0
tox==4.5.2
python-jenkins==1.8.0
requests==2.31.0
xmltodict==0.13.0
